refactor and lats bug fixes
